Refactor(jira): Move the IssueTrackerChannel into alert-common
