+++ alert-common/build.gradle
@@ -1,6 +1,7 @@
 ext.moduleName = 'com.synopsys.integration.alert.common'
 
 dependencies {
+    implementation project(':alert-issuetracker')
     api 'com.blackducksoftware.integration:blackduck-common:44.4.6'
 
     api 'org.springframework.boot:spring-boot-starter-activemq'
