+++ alert-issuetracker/src/test/java/com/synopsys/integration/alert/issuetracker/jira/cloud/JiraCloudIssueConfigValidatorTest.java
@@ -10,7 +10,6 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-import com.synopsys.integration.alert.common.persistence.model.ConfigurationFieldModel;
 import com.synopsys.integration.alert.issuetracker.config.IssueConfig;
 import com.synopsys.integration.alert.issuetracker.config.IssueTrackerContext;
 import com.synopsys.integration.alert.issuetracker.exception.IssueTrackerFieldException;
@@ -35,21 +34,10 @@ public class JiraCloudIssueConfigValidatorTest {
         Mockito.when(issueMetaDataService.doesProjectContainIssueType(Mockito.anyString(), Mockito.anyString())).thenReturn(Boolean.TRUE);
         JiraCloudIssueConfigValidator jiraIssueConfigValidator = new JiraCloudIssueConfigValidator(projectService, userSearchService, issueTypeService, issueMetaDataService);
 
-        ConfigurationFieldModel resolveTransition = ConfigurationFieldModel.create(JiraCloudProperties.KEY_RESOLVE_WORKFLOW_TRANSITION);
         String resolveTransitionString = "Resolve";
-        resolveTransition.setFieldValue(resolveTransitionString);
-
-        ConfigurationFieldModel project = ConfigurationFieldModel.create(JiraCloudProperties.KEY_JIRA_PROJECT_NAME);
         String projectName = "ProjectName";
-        project.setFieldValue(projectName);
-
-        ConfigurationFieldModel issueType = ConfigurationFieldModel.create(JiraCloudProperties.KEY_ISSUE_TYPE);
         String issueTypeString = "IssueType";
-        issueType.setFieldValue(issueTypeString);
-
-        ConfigurationFieldModel issueCreator = ConfigurationFieldModel.create(JiraCloudProperties.KEY_ISSUE_CREATOR);
         String issueCreatorString = "IssueCreator";
-        issueCreator.setFieldValue(issueCreatorString);
 
         IssueConfig issueConfig = new IssueConfig();
         issueConfig.setCommentOnIssues(true);
@@ -92,13 +80,8 @@ public class JiraCloudIssueConfigValidatorTest {
         IssueMetaDataService issueMetaDataService = Mockito.mock(IssueMetaDataService.class);
         JiraCloudIssueConfigValidator jiraIssueConfigValidator = new JiraCloudIssueConfigValidator(projectService, userSearchService, issueTypeService, issueMetaDataService);
 
-        ConfigurationFieldModel resolveTransition = ConfigurationFieldModel.create(JiraCloudProperties.KEY_RESOLVE_WORKFLOW_TRANSITION);
         String resolveTransitionString = "Resolve";
-        resolveTransition.setFieldValue(resolveTransitionString);
-
-        ConfigurationFieldModel issueType = ConfigurationFieldModel.create(JiraCloudProperties.KEY_ISSUE_TYPE);
         String issueTypeString = "IssueType";
-        issueType.setFieldValue(issueTypeString);
 
         IssueConfig issueConfig = new IssueConfig();
         issueConfig.setCommentOnIssues(true);
